+++ server/models/files.js
@@ -36,177 +36,13 @@ module.exports = {
 
     return file
   },
-  get (page) {
-    const files = db.app.prepare(
-      `SELECT
-        id,
-        mime,
-        size,
-        width,
-        height,
-        hash
-      FROM
-        hydrusrv_files
-      LIMIT
-        ${config.resultsPerPage}
-      OFFSET
-        ${(page - 1) * config.resultsPerPage}`
-    ).all()
+  get (page, sort = 'id', namespaces) {
+    const orderBy = this.generateOrderBy(sort, namespaces)
 
-    if (files.length) {
-      for (const file of files) {
-        file.mime = hydrusConfig.availableMimeTypes[file.mime]
-        file.mediaUrl = generateFilePath('original', file.hash)
-        file.thumbnailUrl = generateFilePath('thumbnail', file.hash)
-        delete file.hash
-      }
-    }
-
-    return files
-  },
-  getByTags (page, tags) {
-    tags = [...new Set(tags)]
-
-    const files = db.app.prepare(
-      `SELECT
-        hydrusrv_files.id,
-        hydrusrv_files.mime,
-        hydrusrv_files.size,
-        hydrusrv_files.width,
-        hydrusrv_files.height,
-        hydrusrv_files.hash
-      FROM
-        hydrusrv_files
-      LEFT JOIN
-        hydrusrv_mappings
-        ON hydrusrv_mappings.file_id = hydrusrv_files.id
-      LEFT JOIN
-        hydrusrv_tags
-        ON hydrusrv_tags.id = hydrusrv_mappings.tag_id
-      WHERE
-        hydrusrv_tags.name IN (${',?'.repeat(tags.length).replace(',', '')})
-      GROUP BY
-        hydrusrv_files.id
-      HAVING
-        count(DISTINCT hydrusrv_tags.id) = ?
-      LIMIT
-        ${config.resultsPerPage}
-      OFFSET
-        ${(page - 1) * config.resultsPerPage}`
-    ).all(tags, tags.length)
-
-    if (files.length) {
-      for (const file of files) {
-        file.mime = hydrusConfig.availableMimeTypes[file.mime]
-        file.mediaUrl = generateFilePath('original', file.hash)
-        file.thumbnailUrl = generateFilePath('thumbnail', file.hash)
-        delete file.hash
-      }
-    }
-
-    return files
-  },
-  getSortedRandomly (page) {
-    const files = db.app.prepare(
-      `SELECT
-        id,
-        mime,
-        size,
-        width,
-        height,
-        hash
-      FROM
-        hydrusrv_files
-      ORDER BY
-        random
-      LIMIT
-        ${config.resultsPerPage}
-      OFFSET
-        ${(page - 1) * config.resultsPerPage}`
-    ).all()
-
-    if (files.length) {
-      for (const file of files) {
-        file.mime = hydrusConfig.availableMimeTypes[file.mime]
-        file.mediaUrl = generateFilePath('original', file.hash)
-        file.thumbnailUrl = generateFilePath('thumbnail', file.hash)
-        delete file.hash
-      }
-    }
-
-    return files
-  },
-  getByTagsSortedRandomly (page, tags) {
-    tags = [...new Set(tags)]
-
-    const files = db.app.prepare(
-      `SELECT
-        hydrusrv_files.id,
-        hydrusrv_files.mime,
-        hydrusrv_files.size,
-        hydrusrv_files.width,
-        hydrusrv_files.height,
-        hydrusrv_files.hash
-      FROM
-        hydrusrv_files
-      LEFT JOIN
-        hydrusrv_mappings
-        ON hydrusrv_mappings.file_id = hydrusrv_files.id
-      LEFT JOIN
-        hydrusrv_tags
-        ON hydrusrv_tags.id = hydrusrv_mappings.tag_id
-      WHERE
-        hydrusrv_tags.name IN (${',?'.repeat(tags.length).replace(',', '')})
-      GROUP BY
-        hydrusrv_files.id
-      HAVING
-        count(DISTINCT hydrusrv_tags.id) = ?
-      ORDER BY
-        hydrusrv_files.random
-      LIMIT
-        ${config.resultsPerPage}
-      OFFSET
-        ${(page - 1) * config.resultsPerPage}`
-    ).all(tags, tags.length)
-
-    if (files.length) {
-      for (const file of files) {
-        file.mime = hydrusConfig.availableMimeTypes[file.mime]
-        file.mediaUrl = generateFilePath('original', file.hash)
-        file.thumbnailUrl = generateFilePath('thumbnail', file.hash)
-        delete file.hash
-      }
-    }
-
-    return files
-  },
-  getSortedByNamespaces (page, namespaces) {
-    namespaces = [...new Set(namespaces)]
-
-    const validNamespaces = tagsModel.getNamespaces().map(
-      namespace => namespace.name
-    )
-
-    namespaces = namespaces.filter(
-      namespace => validNamespaces.includes(namespace)
-    )
-
-    if (!namespaces.length) {
+    if (!orderBy) {
       return this.get(page)
     }
 
-    const namespaceOrderBys = []
-
-    for (const namespace of namespaces) {
-      namespaceOrderBys.push(
-        `CASE
-          WHEN namespace_${namespace.split(' ').join('_')} IS NULL THEN 1
-          ELSE 0
-        END,
-        namespace_${namespace.split(' ').join('_')}`
-      )
-    }
-
     const files = db.app.prepare(
       `SELECT
         id,
@@ -218,7 +54,7 @@ module.exports = {
       FROM
         hydrusrv_files
       ORDER BY
-        ${namespaceOrderBys.join(',')}, id
+        ${orderBy}
       LIMIT
         ${config.resultsPerPage}
       OFFSET
@@ -236,32 +72,13 @@ module.exports = {
 
     return files
   },
-  getByTagsSortedByNamespaces (page, tags, namespaces) {
+  getByTags (page, tags, sort = 'id', namespaces) {
     tags = [...new Set(tags)]
-    namespaces = [...new Set(namespaces)]
 
-    const validNamespaces = tagsModel.getNamespaces().map(
-      namespace => namespace.name
-    )
+    const orderBy = this.generateOrderBy(sort, namespaces)
 
-    namespaces = namespaces.filter(
-      namespace => validNamespaces.includes(namespace)
-    )
-
-    if (!namespaces.length) {
+    if (!orderBy) {
       return this.getByTags(page, tags)
-    }
-
-    const namespaceOrderBys = []
-
-    for (const namespace of namespaces) {
-      namespaceOrderBys.push(
-        `CASE
-          WHEN namespace_${namespace.split(' ').join('_')} IS NULL THEN 1
-          ELSE 0
-        END,
-        namespace_${namespace.split(' ').join('_')}`
-      )
     }
 
     const files = db.app.prepare(
@@ -287,7 +104,7 @@ module.exports = {
       HAVING
         count(DISTINCT hydrusrv_tags.id) = ?
       ORDER BY
-        ${namespaceOrderBys.join(',')}, hydrusrv_files.id
+        ${orderBy}
       LIMIT
         ${config.resultsPerPage}
       OFFSET
@@ -309,5 +126,58 @@ module.exports = {
     return db.app.prepare(
       'SELECT COUNT(id) as count FROM hydrusrv_files'
     ).get()
+  },
+  generateOrderBy (sort, namespaces) {
+    if (sort === 'namespace' && namespaces.length) {
+      const namespaceOrderBy = this.generateNamespaceOrderBy(namespaces)
+
+      if (!namespaceOrderBy.length) {
+        return undefined
+      }
+
+      return `${namespaceOrderBy.join(',')}, hydrusrv_files.id`
+    }
+
+    switch (sort) {
+      case 'size':
+        return 'hydrusrv_files.size DESC'
+      case 'width':
+        return 'hydrusrv_files.width DESC'
+      case 'height':
+        return 'hydrusrv_files.height DESC'
+      case 'random':
+        return 'hydrusrv_files.random ASC'
+      default:
+        return 'hydrusrv_files.id ASC'
+    }
+  },
+  generateNamespaceOrderBy (namespaces) {
+    namespaces = [...new Set(namespaces)]
+
+    const validNamespaces = tagsModel.getNamespaces().map(
+      namespace => namespace.name
+    )
+
+    namespaces = namespaces.filter(
+      namespace => validNamespaces.includes(namespace)
+    )
+
+    if (!namespaces.length) {
+      return []
+    }
+
+    const namespaceOrderBy = []
+
+    for (const namespace of namespaces) {
+      namespaceOrderBy.push(
+        `CASE
+          WHEN namespace_${namespace.split(' ').join('_')} IS NULL THEN 1
+          ELSE 0
+        END,
+        namespace_${namespace.split(' ').join('_')}`
+      )
+    }
+
+    return namespaceOrderBy
   }
 }